pyproject.toml
[tool.poetry]
name = "DelePwn"
version = "0.1.0"
description = ""
authors = ["n0tspam <admin@spamsec.pro>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.7"
google-cloud-iam = "^2.12.0"
google-cloud-resource-manager = "^1.10.1"
google-auth-oauthlib = "^1.0.0"
httplib2 = "^0.22.0"
uritemplate = "^4.1.1"
google-auth-httplib2 = "^0.1.0"
pyyaml = "^6.0.1"
google-api-python-client = "^2.95.0"
colorama = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
